% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{widecolorgamut}
\alias{widecolorgamut}
\title{Wide Color Gamut EXR Data}
\format{
An array of four channels (RBGA) and a width/height
}
\source{
<https://openexr.com/en/latest/test_images/TestImages/WideColorGamut.html>
}
\usage{
widecolorgamut
}
\description{
Wide Color Gamut numeric data in RGBA list format from the OpenEXR project.
}
\keyword{datasets}
